+++ nodejs/systemHandler.js
@@ -60,12 +60,12 @@ class SystemHandler {
         logger.fine('Checking db variables.');
         return handleDbVars.call(this)
             .then(() => {
-                logger.fine('Checking NTP.');
-                return handleNTP.call(this);
-            })
-            .then(() => {
                 logger.fine('Checking DNS.');
                 return handleDNS.call(this);
+            })
+            .then(() => {
+                logger.fine('Checking NTP.');
+                return handleNTP.call(this);
             })
             .then(() => {
                 logger.fine('Checking hostname.');
@@ -104,27 +104,40 @@ function handleDbVars() {
 function handleNTP() {
     if (this.declaration.Common.NTP) {
         const ntp = this.declaration.Common.NTP;
-        return this.bigIp.replace(
-            PATHS.NTP,
-            {
-                servers: ntp.servers,
-                timezone: ntp.timezone
-            }
-        );
+        const promises = (ntp.servers || []).map((server) => {
+            return doUtil.checkDnsResolution(server);
+        });
+
+        return Promise.all(promises)
+            .then(() => {
+                return disableDhcpOption.call(this, 'ntp-servers');
+            })
+            .then(() => {
+                this.bigIp.replace(
+                    PATHS.NTP,
+                    {
+                        servers: ntp.servers,
+                        timezone: ntp.timezone
+                    }
+                );
+            });
     }
     return Promise.resolve();
 }
 
 function handleDNS() {
     if (this.declaration.Common.DNS) {
-        const dns = this.declaration.Common.DNS;
-        return this.bigIp.replace(
-            PATHS.DNS,
-            {
-                'name-servers': dns.nameServers,
-                search: dns.search || []
-            }
-        );
+        return disableDhcpOption.call(this, 'domain-name-servers')
+            .then(() => {
+                const dns = this.declaration.Common.DNS;
+                return this.bigIp.replace(
+                    PATHS.DNS,
+                    {
+                        'name-servers': dns.nameServers,
+                        search: dns.search || []
+                    }
+                );
+            });
     }
     return Promise.resolve();
 }
@@ -221,7 +234,8 @@ function handleLicensePool(license) {
         getBigIp = Promise.resolve(this.bigIp);
     }
 
-    return getBigIp
+    return doUtil.checkDnsResolution(license.bigIqHost)
+        .then(() => { return getBigIp; })
         .then((resolvedBigIp) => {
             bigIp = resolvedBigIp;
             let possiblyRevoke;
@@ -301,6 +315,9 @@ function handleLicensePool(license) {
                 return Promise.resolve();
             }
             return this.bigIp.active();
+        })
+        .catch((err) => {
+            return Promise.reject(err);
         });
 }
 
@@ -376,6 +393,48 @@ function createOrUpdateUser(username, data) {
         .catch((err) => {
             logger.severe(`Error creating/updating user: ${err.message}`);
             return Promise.reject(err);
+        });
+}
+
+function disableDhcpOption(optionToDisable) {
+    let requiresDhcpRestart;
+
+    return this.bigIp.list(
+        '/tm/sys/management-dhcp/sys-mgmt-dhcp-config'
+    )
+        .then((dhcpOptions) => {
+            if (!dhcpOptions || !dhcpOptions.requestOptions) {
+                return Promise.resolve();
+            }
+
+            const currentOptions = dhcpOptions.requestOptions;
+            const newOptions = currentOptions.filter((option) => {
+                return option !== optionToDisable;
+            });
+
+            if (currentOptions.length === newOptions.length) {
+                return Promise.resolve();
+            }
+
+            requiresDhcpRestart = true;
+            return this.bigIp.modify(
+                '/tm/sys/management-dhcp/sys-mgmt-dhcp-config',
+                {
+                    requestOptions: newOptions
+                }
+            );
+        })
+        .then(() => {
+            if (!requiresDhcpRestart) {
+                return Promise.resolve();
+            }
+            return this.bigIp.create(
+                '/tm/sys/service',
+                {
+                    command: 'restart',
+                    name: 'dhclient'
+                }
+            );
         });
 }