+++ test/nodejs/restWorkerTests.js
@@ -18,6 +18,7 @@
 
 const assert = require('assert');
 const State = require('../../nodejs/state');
+const RealValidator = require('../../nodejs/validator');
 const STATUS = require('../../nodejs/sharedConstants').STATUS;
 const EVENTS = require('../../nodejs/sharedConstants').EVENTS;
 
@@ -71,7 +72,13 @@ describe('restWorker', () => {
     });
 
     describe('onStart', () => {
+        let restWorker;
         beforeEach(() => {
+            restWorker = new RestWorker();
+            restWorker.restHelper = {
+                makeRestjavadUri() {}
+            };
+            restWorker.dependencies = [];
         });
 
         it('should respond handle success', () => {
@@ -83,7 +90,6 @@ describe('restWorker', () => {
                     reject(new Error('should have called success'));
                 };
 
-                const restWorker = new RestWorker();
                 try {
                     restWorker.onStart(success, error);
                 } catch (err) {
@@ -101,7 +107,29 @@ describe('restWorker', () => {
                     resolve();
                 };
 
-                const restWorker = new RestWorker();
+                try {
+                    restWorker.onStart(success, error);
+                } catch (err) {
+                    reject(err);
+                }
+            });
+        });
+
+        it('should set dependencies', () => {
+            return new Promise((resolve, reject) => {
+                const deviceInfoUri = 'https://path/to/deviceInfo';
+                restWorker.restHelper.makeRestjavadUri = () => {
+                    return deviceInfoUri;
+                };
+
+                const success = () => {
+                    assert.strictEqual(restWorker.dependencies[0], deviceInfoUri);
+                    resolve();
+                };
+                const error = () => {
+                    reject(new Error('should have called success'));
+                };
+
                 try {
                     restWorker.onStart(success, error);
                 } catch (err) {
@@ -382,12 +410,14 @@ describe('restWorker', () => {
                         assert.strictEqual(
                             restWorker
                                 .state.doState.tasks[1234].internalDeclaration.Common.myLicense.bigIpPassword,
-                            bigIpPassword
+                            undefined,
+                            'Stored state should not save decryptedId'
                         );
                         assert.strictEqual(
                             restWorker
                                 .state.doState.tasks[1234].internalDeclaration.Common.myLicense.bigIqPassword,
-                            bigIqPassword
+                            undefined,
+                            'Stored state should not save decryptedId'
                         );
                         resolve();
                     };
@@ -693,9 +723,9 @@ describe('restWorker', () => {
             };
 
             validatorMock.validate = () => {
-                return {
+                return Promise.resolve({
                     isValid: true
-                };
+                });
             };
 
             RestWorker.prototype.saveState = (foo, state, callback) => {
@@ -769,9 +799,9 @@ describe('restWorker', () => {
         it('should handle validation errors', () => {
             return new Promise((resolve, reject) => {
                 validatorMock.validate = () => {
-                    return {
+                    return Promise.resolve({
                         isValid: false
-                    };
+                    });
                 };
 
                 restOperationMock.complete = () => {
@@ -817,6 +847,40 @@ describe('restWorker', () => {
 
                 restOperationMock.getContentType = () => {};
                 restOperationMock.getBody = () => { return 'foobar'; };
+
+                try {
+                    restWorker.onPost(restOperationMock);
+                } catch (err) {
+                    reject(err);
+                }
+            });
+        });
+
+        it('should add defaults to the declaration', () => {
+            return new Promise((resolve, reject) => {
+                declaration = {
+                    schemaVersion: '1.0.0',
+                    class: 'Device',
+                    Common: {
+                        mySelfIp: {
+                            class: 'SelfIp',
+                            address: '1.2.3.4',
+                            vlan: 'foo'
+                        }
+                    }
+
+                };
+                restWorker.validator = new RealValidator();
+
+                restOperationMock.complete = () => {
+                    const state = restWorker.state.doState;
+                    const taskId = Object.keys(state.tasks)[0];
+                    assert.strictEqual(
+                        state.tasks[taskId].internalDeclaration.Common.mySelfIp.trafficGroup,
+                        'traffic-group-local-only'
+                    );
+                    resolve();
+                };
 
                 try {
                     restWorker.onPost(restOperationMock);